backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
groq==0.4.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
tensorflow==2.14.0
pypdf2==3.0.1
python-docx==1.0.1
aiofiles==23.2.0
httpx==0.25.2